requirements.txt
pandas==1.0.3
requests==2.23.0
urllib3==1.25.8
python_upwork==1.3.6
BeautifulSoup4==4.9.0
lxml==4.5.0